Make sure bind mount paths are relative to the app host directory (#3467)
* Make sure bind mount paths are relative to the app host directory
- Made it so that bind mounts annotations are resolved relative to the app host directory by default.
- Bind mounts without a rooted path will throw an error.
- Removed logic from ApplicationExecutor that checks the annotation data since this is checked in the constructor.
- Added tests (moved tests to Containers folder)

* Fixed tests and move more tests out of ManifestGenerationTests class
