+++ src/Aspire.Hosting/DistributedApplicationOptions.cs
@@ -11,7 +11,9 @@ namespace Aspire.Hosting;
 public sealed class DistributedApplicationOptions
 {
     private readonly Lazy<Assembly?> _assembly;
-    private readonly Lazy<string?> _projectDirectory;
+    private readonly Lazy<string?> _projectDirectoryLazy;
+    // This is for testing
+    private string? _projectDirectory;
 
     /// <summary>
     /// Initializes a new instance of <see cref="DistributedApplicationOptions"/>.
@@ -19,7 +21,7 @@ public sealed class DistributedApplicationOptions
     public DistributedApplicationOptions()
     {
         _assembly = new(ResolveAssembly);
-        _projectDirectory = new(ResolveProjectDirectory);
+        _projectDirectoryLazy = new(ResolveProjectDirectory);
     }
 
     /// <summary>
@@ -39,7 +41,11 @@ public sealed class DistributedApplicationOptions
 
     internal Assembly? Assembly => _assembly.Value;
 
-    internal string? ProjectDirectory => _projectDirectory.Value;
+    internal string? ProjectDirectory
+    {
+        get => _projectDirectory ?? _projectDirectoryLazy.Value;
+        set => _projectDirectory = value;
+    }
 
     internal bool DashboardEnabled => !DisableDashboard;
