+++ src/Aspire.Hosting/Properties/xlf/Resources.es.xlf
@@ -7,6 +7,11 @@
         <target state="translated">Los volúmenes anónimos no pueden ser de solo lectura.</target>
         <note />
       </trans-unit>
+      <trans-unit id="ContainerMountBindMountsRequireRootedPaths">
+        <source>Bind mounts must specify an absolute path.</source>
+        <target state="new">Bind mounts must specify an absolute path.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="ContainerMountBindMountsRequireSourceExceptionMessage">
         <source>Bind mounts must specify a source path.</source>
         <target state="translated">Los montajes de enlace deben especificar una ruta de acceso de origen.</target>
