+++ src/Aspire.Hosting/Properties/xlf/Resources.pl.xlf
@@ -7,6 +7,11 @@
         <target state="translated">Woluminy anonimowe nie mogą być tylko do odczytu.</target>
         <note />
       </trans-unit>
+      <trans-unit id="ContainerMountBindMountsRequireRootedPaths">
+        <source>Bind mounts must specify an absolute path.</source>
+        <target state="new">Bind mounts must specify an absolute path.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="ContainerMountBindMountsRequireSourceExceptionMessage">
         <source>Bind mounts must specify a source path.</source>
         <target state="translated">Instalacje wiązania muszą określać ścieżkę źródłową.</target>
