+++ src/Aspire.Hosting/Properties/xlf/Resources.pt-BR.xlf
@@ -7,6 +7,11 @@
         <target state="translated">Volumes anônimos não podem ser somente leitura.</target>
         <note />
       </trans-unit>
+      <trans-unit id="ContainerMountBindMountsRequireRootedPaths">
+        <source>Bind mounts must specify an absolute path.</source>
+        <target state="new">Bind mounts must specify an absolute path.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="ContainerMountBindMountsRequireSourceExceptionMessage">
         <source>Bind mounts must specify a source path.</source>
         <target state="translated">As montagens de associação devem especificar um caminho de origem.</target>
