+++ src/Aspire.Hosting/Properties/xlf/Resources.zh-Hans.xlf
@@ -7,6 +7,11 @@
         <target state="translated">匿名卷不能为只读。</target>
         <note />
       </trans-unit>
+      <trans-unit id="ContainerMountBindMountsRequireRootedPaths">
+        <source>Bind mounts must specify an absolute path.</source>
+        <target state="new">Bind mounts must specify an absolute path.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="ContainerMountBindMountsRequireSourceExceptionMessage">
         <source>Bind mounts must specify a source path.</source>
         <target state="translated">绑定装载必须指定源路径。</target>
