+++ src/Aspire.Hosting/Properties/xlf/Resources.zh-Hant.xlf
@@ -7,6 +7,11 @@
         <target state="translated">匿名磁碟區不可為唯讀。</target>
         <note />
       </trans-unit>
+      <trans-unit id="ContainerMountBindMountsRequireRootedPaths">
+        <source>Bind mounts must specify an absolute path.</source>
+        <target state="new">Bind mounts must specify an absolute path.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="ContainerMountBindMountsRequireSourceExceptionMessage">
         <source>Bind mounts must specify a source path.</source>
         <target state="translated">繫結裝載必須指定來源路徑。</target>
