+++ tests/Aspire.Hosting.Tests/Containers/ContainerImageAnnotationTests.cs
