+++ tests/Aspire.Hosting.Tests/Containers/ContainerResourceBuilderTests.cs
