+++ tests/Aspire.Hosting.Tests/ManifestGenerationTests.cs
@@ -136,140 +136,6 @@ public class ManifestGenerationTests
             "Service has no bindings because they weren't populated from the launch profile.");
     }
 
-    [Fact]
-    public void EnsureContainerWithEndpointsEmitsContainerPort()
-    {
-        using var program = CreateTestProgramJsonDocumentManifestPublisher();
-
-        program.AppBuilder.AddContainer("grafana", "grafana/grafana")
-                          .WithHttpEndpoint(3000);
-
-        // Build AppHost so that publisher can be resolved.
-        program.Build();
-        var publisher = program.GetManifestPublisher();
-
-        program.Run();
-
-        var resources = publisher.ManifestDocument.RootElement.GetProperty("resources");
-
-        var grafana = resources.GetProperty("grafana");
-        var bindings = grafana.GetProperty("bindings");
-        var httpBinding = bindings.GetProperty("http");
-        Assert.Equal(3000, httpBinding.GetProperty("targetPort").GetInt32());
-    }
-
-    [Fact]
-    public void EnsureContainerWithArgsEmitsContainerArgs()
-    {
-        using var program = CreateTestProgramJsonDocumentManifestPublisher();
-
-        program.AppBuilder.AddContainer("grafana", "grafana/grafana")
-                          .WithArgs("test", "arg2", "more");
-
-        // Build AppHost so that publisher can be resolved.
-        program.Build();
-        var publisher = program.GetManifestPublisher();
-
-        program.Run();
-
-        var resources = publisher.ManifestDocument.RootElement.GetProperty("resources");
-
-        var grafana = resources.GetProperty("grafana");
-        var args = grafana.GetProperty("args");
-        Assert.Equal(3, args.GetArrayLength());
-        Assert.Collection(args.EnumerateArray(),
-            arg => Assert.Equal("test", arg.GetString()),
-            arg => Assert.Equal("arg2", arg.GetString()),
-            arg => Assert.Equal("more", arg.GetString()));
-    }
-
-    [Fact]
-    public async Task EnsureContainerWithVolumesEmitsVolumes()
-    {
-        using var program = CreateTestProgramJsonDocumentManifestPublisher();
-
-        var container = program.AppBuilder.AddContainer("containerwithvolumes", "image/name")
-            .WithVolume("myvolume", "/mount/here")
-            .WithVolume("myreadonlyvolume", "/mount/there", isReadOnly: true)
-            .WithVolume(null! /* anonymous volume */, "/mount/everywhere");
-
-        program.Build();
-
-        var manifest = await ManifestUtils.GetManifest(container.Resource);
-
-        var expectedManifest = """
-            {
-              "type": "container.v0",
-              "image": "image/name:latest",
-              "volumes": [
-                {
-                  "name": "myvolume",
-                  "target": "/mount/here",
-                  "readOnly": false
-                },
-                {
-                  "name": "myreadonlyvolume",
-                  "target": "/mount/there",
-                  "readOnly": true
-                },
-                {
-                  "target": "/mount/everywhere",
-                  "readOnly": false
-                }
-              ]
-            }
-            """;
-
-        Assert.Equal(expectedManifest, manifest.ToString());
-    }
-
-    [Fact]
-    public async Task EnsureContainerWithBindMountsEmitsBindMounts()
-    {
-        using var program = CreateTestProgramJsonDocumentManifestPublisher();
-
-        var container = program.AppBuilder.AddContainer("containerwithbindmounts", "image/name")
-            .WithBindMount("./some/source", "/bound")
-            .WithBindMount("not/relative/qualified", "/another/place")
-            .WithBindMount(".\\some\\other\\source", "\\mount\\here")
-            .WithBindMount("./some/file/path.txt", "/mount/there.txt", isReadOnly: true);
-
-        program.Build();
-
-        var manifest = await ManifestUtils.GetManifest(container.Resource);
-
-        var expectedManifest = """
-            {
-              "type": "container.v0",
-              "image": "image/name:latest",
-              "bindMounts": [
-                {
-                  "source": "some/source",
-                  "target": "/bound",
-                  "readOnly": false
-                },
-                {
-                  "source": "not/relative/qualified",
-                  "target": "/another/place",
-                  "readOnly": false
-                },
-                {
-                  "source": "some/other/source",
-                  "target": "/mount/here",
-                  "readOnly": false
-                },
-                {
-                  "source": "some/file/path.txt",
-                  "target": "/mount/there.txt",
-                  "readOnly": true
-                }
-              ]
-            }
-            """;
-
-        Assert.Equal(expectedManifest, manifest.ToString());
-    }
-
     [Theory]
     [InlineData(new string[] { "args1", "args2" }, new string[] { "withArgs1", "withArgs2" })]
     [InlineData(new string[] { }, new string[] { "withArgs1", "withArgs2" })]
@@ -327,27 +193,6 @@ public class ManifestGenerationTests
         var resource = resources.GetProperty("program");
         var exists = resource.TryGetProperty("args", out _);
         Assert.False(exists);
-    }
-
-    [Fact]
-    public void EnsureContainerWithCustomEntrypointEmitsEntrypoint()
-    {
-        using var program = CreateTestProgramJsonDocumentManifestPublisher();
-
-        var container = program.AppBuilder.AddContainer("grafana", "grafana/grafana");
-        container.Resource.Entrypoint = "custom";
-
-        // Build AppHost so that publisher can be resolved.
-        program.Build();
-        var publisher = program.GetManifestPublisher();
-
-        program.Run();
-
-        var resources = publisher.ManifestDocument.RootElement.GetProperty("resources");
-
-        var grafana = resources.GetProperty("grafana");
-        var entrypoint = grafana.GetProperty("entrypoint");
-        Assert.Equal("custom", entrypoint.GetString());
     }
 
     [Fact]
