+++ tests/Aspire.Hosting.Tests/Redis/AddRedisTests.cs
@@ -213,7 +213,7 @@ public class AddRedisTests
 
         var volumeAnnotation = redis.Resource.Annotations.OfType<ContainerMountAnnotation>().Single();
 
-        Assert.Equal("mydata", volumeAnnotation.Source);
+        Assert.Equal(Path.GetFullPath("mydata"), volumeAnnotation.Source);
         Assert.Equal("/data", volumeAnnotation.Target);
         Assert.Equal(ContainerMountType.BindMount, volumeAnnotation.Type);
         Assert.Equal(isReadOnly ?? false, volumeAnnotation.IsReadOnly);
