+++ tests/Aspire.Hosting.Tests/Utils/ManifestUtils.cs
@@ -10,13 +10,15 @@ namespace Aspire.Hosting.Utils;
 
 internal sealed class ManifestUtils
 {
-    public static async Task<JsonNode> GetManifest(IResource resource)
+    public static async Task<JsonNode> GetManifest(IResource resource, string? manifestDirectory = null)
     {
+        manifestDirectory ??= Environment.CurrentDirectory;
+
         using var ms = new MemoryStream();
         var writer = new Utf8JsonWriter(ms);
         var executionContext = new DistributedApplicationExecutionContext(DistributedApplicationOperation.Publish);
         writer.WriteStartObject();
-        var context = new ManifestPublishingContext(executionContext, Path.Combine(Environment.CurrentDirectory, "manifest.json"), writer);
+        var context = new ManifestPublishingContext(executionContext, Path.Combine(manifestDirectory, "manifest.json"), writer);
         await context.WriteResourceAsync(resource);
         writer.WriteEndObject();
         writer.Flush();
